+++ src/conformly/_internal/api/case.py
@@ -1,7 +1,7 @@
 from typing import Any
 
 from ._errors import api_error
-from ._utils import ensure_model_or_spec, plan_tasks
+from ._utils import ensure_model_or_spec, normalize_overrides, plan_tasks
 from .path import PathSelector
 
 from conformly._internal.generator.context import create_context
@@ -16,6 +16,7 @@ def case(
     valid: bool = True,
     seed: int | None = None,
     strategy: CaseStrategy | PathSelector = "first",
+    overrides: list[PathSelector] | None = None,
     allow_type_mismatch: bool = False,
 ) -> dict[str, Any]:
     """
@@ -58,6 +59,12 @@ def case(
             - PathSelector DSL:
                 path("user.email").violate(V.TOO_SHORT)
 
+        overrides:
+            Optional list of `PathSelector` to set field values.
+            - Applied after generation
+            - Override generated values
+            - With valid=False: act as defaults (ignored if field is violated)
+
         allow_type_mismatch:
             If True fields could be type mismatched.
             Availiable only when valid=False.
@@ -68,6 +75,8 @@ def case(
     model = ensure_model_or_spec(model_or_spec)
 
     ctx = create_context(seed)
+
+    _overrides = normalize_overrides(model, overrides)
 
     if valid:
         if allow_type_mismatch:
@@ -84,7 +93,7 @@ def case(
                 strategy=strategy,
             )
 
-        return generate_valid(ctx, model)
+        return generate_valid(ctx, model, _overrides)
 
     if strategy == "all":
         raise api_error(
@@ -102,4 +111,4 @@ def case(
         count=1,
         allow_type_mismatch=allow_type_mismatch,
     )[0]
-    return generate_invalid(ctx, model, task)
+    return generate_invalid(ctx, model, task, _overrides)