+++ CHANGELOG.md
@@ -5,6 +5,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.7.0] - 2026.06.20
+
+### Added
+- **Overrides support** in `case()` and `cases()` to explicitly set field values
+    - Syntax example `path(T, lambda t: t.name).set("name")`
+    - Overrides has higher priotity than field default
+
+### Changed
+- Now `path` uses proxy-object instead of ast parsing
+
+
 ## [0.6.0] - 2026.06.13
 
 ### Added